removing useless code and add tests
Removing code to dynamically load libraries, since it will not be really implemented right now, it will just be a pain to maintain.

Adding tests to check printing in all three cases of tests (success, fail, and skip). But for now, suites aren't really tested, and data aren't tested either.

+++ include/internal/printer.h
@@ -6,7 +6,7 @@ extern "C" {
 #endif
 
 #include <stdint.h>
-
+#include <stdlib.h>
 #include <stdbool.h>
 
 /**
@@ -26,7 +26,7 @@ typedef struct clarity_test_result_s {
 	/**
 	 * @brief The line number where the test was defined.
 	 */
-	int line_number;
+	size_t line_number;
 
 	/**
 	 * @brief Indicates if the test passed or failed.
